make reuseable func
